+++ spotbugsTestCases/src/java/ghIssues/Issue872.java
@@ -45,6 +45,39 @@ public class Issue872 {
 		Mockito.verify(checked, Mockito.times(4)).noSideEffect();
 	}
 	
+	// compliant
+	void doAnswerMockito() {
+		Value spy = Mockito.spy(Value.class);
+		Mockito.doAnswer(invocation -> {
+				return "foo";
+			}
+		).when(spy).checkMe();
+	}
+	
+	// compliant
+	void doReturnMockito() {
+		Value receiver = Mockito.mock(Value.class);
+		Mockito.doReturn("foo").when(receiver).checkMe();
+	}
+	
+	// compliant
+	void doCallRealMethodMockito() {
+		Value receiver = Mockito.mock(Value.class);
+		Mockito.doCallRealMethod().when(receiver).checkMe();
+	}
+	
+	// compliant
+	void doNothingMockito() {
+		Value receiver = Mockito.mock(Value.class);
+		Mockito.doNothing().when(receiver).checkMe();
+	}
+	
+	// compliant
+	void doThrowMockito() {
+		Value receiver = Mockito.mock(Value.class);
+		Mockito.doThrow(Exception.class).when(receiver).checkMe();
+	}
+	
 	public static class Value {
 		@CheckReturnValue
 		public String checkMe() {